Configure CI
* cargo build
* cargo test
* cargo run --example test spec.json

Every CI job saves a cache to reduce build times. Cache is job specific
so test does not use the result of build for example.

This way we need to store caches for each job. On the other hand all
jobs can run in parallel instead of having to wait for build to finish.
Since build is plenty fast we'll need to check how the caches
behave (invalidation time, size) to see what is the best alternative.

Bors was configured has a merge bot, it didn't work so well using the
hosted app and since this is a one man shop for now a merge bot doesn't
seem so necessary.